Create README.md
This project aims to build a financial content curation site that users can visit to get the latest financial news in one place.
Users will also be able to log in to the site where they will be able to add stocks and cryptocurrencies to their basket. 
The basket will display the latest price data for each stock and cryptocurrency inside it. 
Users will be able to click on the stock ticker to visit a detail page where all the meta data and price data for that particular stock or cryptocurrency is displayed.